Filtering taxa who is low relative abundance or unclassified

whether to filter the low relative abundance or unclassified taxa by the threshold. Here, we choose the following criterion:

Taxa more than Mean relative abundance across all samples: 0.0001 (1e-04);

Taxa more than Minimum relative abundance at least one sample: 0.001 (1e-03).

filter_abundance(
    object,
    level = NULL,
    cutoff_mean = c(100, 1e-04),
    cutoff_one = c(1000, 1e-03),
    unclass = TRUE)

Arguments

object

(Required). a phyloseq-class object.

level

(Optional). character. taxonomic level to summarize, default the top level rank of the ps. taxonomic level(Kingdom, Phylum, Class, Order, Family, Genus, Species, Strains; default: NULL).

cutoff_mean

(Optional). numeric. Threshold for Mean relative abundance all samples (default, 1e-04).

cutoff_one

(Optional). numeric. Threshold for Minimum relative abundance at least one sample (default, 1e-03).

unclass

(Optional). logical. whether to filter the unclassified taxa (default TRUE).

Value

a phyloseq::phyloseq object, where each row represents a taxa, and each col represents the taxa abundance of each sample.

References

Thingholm, Louise B., et al. "Obese individuals with and without type 2 diabetes show different gut microbial functional capacity and composition." Cell host & microbe 26.2 (2019): 252-264.

Author

Created by Hua Zou (11/30/2021 Shenzhen China)

Examples

if (FALSE) {
 data("enterotypes_arumugam")

 # absolute abundance
 ps <- filter_abundance(
   object = enterotypes_arumugam,
   level = NULL,
   cutoff_mean = 100,
   cutoff_one = 1000,
   unclass = FALSE)

 # relative abundance
 enterotypes_arumugam_rb <- normalize(enterotypes_arumugam, method = "TSS")
 ps <- filter_abundance(
   object = enterotypes_arumugam_rb,
   level = "Phylum",
   cutoff_mean = 1e-04,
   cutoff_one = 1e-03,
   unclass = TRUE)

}