Analysis of Similarity (ANOSIM)

The run_ANOSIM function is used to identify the association between the community and environmental variables by rank.

run_ANOSIM(object,
                  variable,
                  method,
                  seedNum,
                  alpha)

Arguments

object

(Required). a phyloseq::phyloseq or Biobase::ExpressionSet object.

variable

(Required). character. grouping variable for test.

method

(Optional). character. Provide one of the currently supported options. See distanceMethodList for a detailed list of the supported options and links to accompanying documentation. Options include:

"unifrac" : unweighted UniFrac distance.

"wunifrac": weighted-UniFrac distance.

"GUniFrac": The variance-adjusted weighted UniFrac distances (default: alpha=0.5).

"bray": bray crutis distance.

"dpcoa": sample-wise distance used in Double Principle Coordinate Analysis.

"jsd": Jensen-Shannon Divergence. Alternatively, you can provide a character string that defines a custom distance method, if it has the form described in designdist (default: "bray").

seedNum

(Optional). numeric. specify seeds for reproduction (default: 123).

alpha

(Optional). numeric. the parameter for "GUniFrac" controlling weight on abundant lineages (default: 0.5).

Value

An ANOSIM model.

Details

The run_ANOSIM function is used to identify the association between the community and environmental variables by rank, applying the distance in profile and calculating the R statistic between community and rank variable by permutation test to determine the significance. It can be applied to both phyloseq::phyloseq and Biobase::ExpressionSet object.

References

Clarke, K. Robert. "Non‐parametric multivariate analyses of changes in community structure." Australian journal of ecology 18.1 (1993): 117-143.

Author

Created by Hua Zou (5/15/2022 Shenzhen China)

Examples

if (FALSE) {
data("enterotypes_arumugam")
run_ANOSIM(enterotypes_arumugam, variable = "Enterotype", method = "bray")
}