Group Standard Error Interval

Calculates the standard error interval of grouped data.

run_group.STDERR(x, data)

Arguments

x

(Required). an aggregate compatible formula.

data

(Required). a data frame (or list) from which the variables in formula should be taken.

Value

A data frame consisting of one column for each grouping factor plus three columns for the upper bound, mean and lower bound of the standard error interval for each level of the grouping factor.

References

https://github.com/cran/Rmisc/

Author

Created by Hua Zou (5/19/2022 Shenzhen China)

Examples

run_group.STDERR(Sepal.Length~Species, iris)
#>      Species Sepal.Length.upper Sepal.Length.mean Sepal.Length.lower
#> 1     setosa           5.055850             5.006           4.956150
#> 2 versicolor           6.008998             5.936           5.863002
#> 3  virginica           6.677927             6.588           6.498073